+++ android/src/org/coolreader/crengine/InterfaceTheme.java
@@ -21,95 +21,7 @@
 
 package org.coolreader.crengine;
 
-import org.coolreader.R;
-
-public class InterfaceTheme {
-
-	public final static InterfaceTheme BLACK = new InterfaceTheme("BLACK", R.style.Theme_Black, R.style.Theme_Black_Dialog_Normal, R.style.Theme_Black_Dialog_Fullscreen, R.string.options_app_ui_theme_black, 0xFF000000)
-			.setRootDelimiter(R.drawable.divider_black_tiled, 2)
-			.setBackgrounds(
-					R.drawable.ui_status_background_browser_black, 
-					R.drawable.ui_toolbar_background_browser_black,
-					R.drawable.ui_toolbar_background_browser_vertical_black,
-					0, 0xFF000000)
-			.setToolbarButtonAlpha(DeviceInfo.EINK_SCREEN ? 0xFF : 0x80);
-
-	public final static InterfaceTheme WHITE = new InterfaceTheme("WHITE", R.style.Theme_White, R.style.Theme_White_Dialog_Normal, R.style.Theme_White_Dialog_Fullscreen, R.string.options_app_ui_theme_white, 0xFFFFFFFF)
-			.setRootDelimiter(R.drawable.divider_white_tiled, 2)
-			.setBackgrounds(
-					R.drawable.ui_status_background_browser_white, 
-					R.drawable.ui_toolbar_background_browser_white,
-					R.drawable.ui_toolbar_background_browser_vertical_white,
-					0, 0xFFFFFFFF)
-			.setToolbarButtonAlpha(DeviceInfo.EINK_SCREEN ? 0xFF : 0xE0);
-
-	public final static InterfaceTheme LIGHT = new InterfaceTheme("LIGHT", R.style.Theme_Light, R.style.Theme_Light_Dialog_Normal, R.style.Theme_Light_Dialog_Fullscreen, R.string.options_app_ui_theme_light, 0xFF000000)
-			.setRootDelimiter(R.drawable.divider_light_tiled, 16)
-			.setBackgrounds(
-					R.drawable.ui_status_background_browser_light, 
-					R.drawable.ui_toolbar_background_browser_light,
-					R.drawable.ui_toolbar_background_browser_vertical_light,
-					R.drawable.background_tiled_light, 0)
-			.setToolbarButtonAlpha(DeviceInfo.EINK_SCREEN ? 0xFF : 0xC0);
-//			.setBackgrounds(
-//					R.drawable.ui_status_background_browser_light, 
-//					R.drawable.ui_toolbar_background_browser_light,
-//					R.drawable.ui_toolbar_background_browser_vertical_light,
-//					R.drawable.ui_status_background_reader_light, 
-//					R.drawable.ui_toolbar_background_reader_light,
-//					R.drawable.ui_toolbar_background_reader_vertical_light,
-//					0x404040, 0xC0);
-
-	public final static InterfaceTheme DARK = new InterfaceTheme("DARK", R.style.Theme_Dark, R.style.Theme_Dark_Dialog_Normal, R.style.Theme_Dark_Dialog_Fullscreen, R.string.options_app_ui_theme_dark, 0xFF000000)
-			.setRootDelimiter(R.drawable.divider_dark_tiled, 16)
-			.setBackgrounds(
-					R.drawable.ui_status_background_browser_dark, 
-					R.drawable.ui_toolbar_background_browser_dark,
-					R.drawable.ui_toolbar_background_browser_vertical_dark,
-					R.drawable.background_tiled_dark, 0)
-			.setToolbarButtonAlpha(DeviceInfo.EINK_SCREEN ? 0xFF : 0x90);
-//			.setBackgrounds(
-//					R.drawable.ui_status_background_browser_dark,
-//					R.drawable.ui_toolbar_background_browser_dark,
-//					R.drawable.ui_toolbar_background_browser_vertical_dark,
-//					R.drawable.ui_status_background_reader_dark,
-//					R.drawable.ui_toolbar_background_reader_dark,
-//					R.drawable.ui_toolbar_background_reader_vertical_dark,
-//					0xE0E0E0, 0x90);
-
-	public final static InterfaceTheme GRAY1 = new InterfaceTheme("GRAY1", R.style.Theme_Gray1, R.style.Theme_Gray1_Dialog_Normal, R.style.Theme_Gray1_Dialog_Fullscreen, R.string.options_app_ui_theme_gray1, 0xFF555555)
-			.setRootDelimiter(R.drawable.divider_black_tiled, 2)
-			.setBackgrounds(
-					R.drawable.ui_status_background_browser_gray1,
-					R.drawable.ui_toolbar_background_browser_gray1,
-					R.drawable.ui_toolbar_background_browser_vertical_gray1,
-					0, 0xFF555555)
-			.setToolbarButtonAlpha(DeviceInfo.EINK_SCREEN ? 0xFF : 0x80);
-
-	public final static InterfaceTheme GRAY2 = new InterfaceTheme("GRAY2", R.style.Theme_Gray2, R.style.Theme_Gray2_Dialog_Normal, R.style.Theme_Gray2_Dialog_Fullscreen, R.string.options_app_ui_theme_gray2, 0xFFCCCCCC)
-			.setRootDelimiter(R.drawable.divider_white_tiled, 2)
-			.setBackgrounds(
-					R.drawable.ui_status_background_browser_gray2,
-					R.drawable.ui_toolbar_background_browser_gray2,
-					R.drawable.ui_toolbar_background_browser_vertical_gray2,
-					0, 0xFFCCCCCC)
-			.setToolbarButtonAlpha(DeviceInfo.EINK_SCREEN ? 0xFF : 0xE0);
-	public final static InterfaceTheme HICONTRAST1 = new InterfaceTheme("HICONTRAST1", R.style.Theme_HiContrast1, R.style.Theme_HiContrast1_Dialog_Normal, R.style.Theme_HiContrast1_Dialog_Fullscreen, R.string.options_app_ui_theme_hicontrast1, 0xFFFFFFFF)
-			.setRootDelimiter(R.drawable.divider_white_tiled, 2)
-			.setBackgrounds(
-					R.drawable.ui_status_background_browser_white,
-					R.drawable.ui_toolbar_background_browser_white,
-					R.drawable.ui_toolbar_background_browser_vertical_white,
-					0, 0xFFFFFFFF)
-			.setToolbarButtonAlpha(0xFF);
-	public final static InterfaceTheme HICONTRAST2 = new InterfaceTheme("HICONTRAST2", R.style.Theme_HiContrast2, R.style.Theme_HiContrast2_Dialog_Normal, R.style.Theme_HiContrast2_Dialog_Fullscreen, R.string.options_app_ui_theme_hicontrast2, 0xFFFFFFFF)
-			.setRootDelimiter(R.drawable.divider_black_tiled, 2)
-			.setBackgrounds(
-					R.drawable.ui_status_background_browser_black,
-					R.drawable.ui_toolbar_background_browser_black,
-					R.drawable.ui_toolbar_background_browser_vertical_black,
-					0, 0xFF000000)
-			.setToolbarButtonAlpha(0xFF);
+public final class InterfaceTheme {
 
 	public String getCode() {
 		return code;
@@ -144,19 +56,21 @@ public class InterfaceTheme {
 //	}
 	
 	public int getRootDelimiterResourceId() {
-		return rootDelimiterResourceId;
+		return visuals.rootDelimiterResourceId;
 	}
 	
 	public int getRootDelimiterHeight() {
-		return rootDelimiterHeight;
+		return visuals.rootDelimiterHeight;
 	}
 	
 	public int getBrowserStatusBackground() {
-		return browserStatusBackground;
+		return visuals.browserStatusBackground;
 	}
 	
 	public int getBrowserToolbarBackground(boolean vertical) {
-		return !vertical ? browserToolbarBackground : browserToolbarBackgroundVertical;
+		return !vertical
+				? visuals.browserToolbarBackground
+				: visuals.browserToolbarBackgroundVertical;
 	}
 	
 //	public int getReaderStatusBackground() {
@@ -172,94 +86,73 @@ public class InterfaceTheme {
 //	}
 	
 	public int getToolbarButtonAlpha() {
-		return toolbarButtonAlpha;
+		return visuals.toolbarButtonAlpha;
 	}
 
 	public int getPopupToolbarBackground() {
-		return popupToolbarBackground;
+		return visuals.popupToolbarBackground;
 	}
 
 	public int getPopupToolbarBackgroundColor() {
-		return popupToolbarBackgroundColor;
+		return visuals.popupToolbarBackgroundColor;
 	}
 
-	
-	public final static InterfaceTheme[] allThemes = {
-		BLACK, WHITE, DARK, LIGHT, GRAY1, GRAY2, HICONTRAST1, HICONTRAST2
-	};
-	
-	public static InterfaceTheme findByCode(String code) {
-		if (code == null)
-			return null;
-		for (InterfaceTheme t : allThemes)
-			if (t.getCode().equals(code))
-				return t;
-		return null;
-	}
-	
 	private final String code;
 	private final int themeId;
 	private final int dialogThemeId;
 	private final int fsDialogThemeId;
 	private final int displayNameResourceId;
 	private final int actionBarBackgroundColorReading;
-	private int rootDelimiterResourceId;
-	private int rootDelimiterHeight;
-	
-	private int browserStatusBackground;
-	private int browserToolbarBackground;
-	private int browserToolbarBackgroundVertical;
-	private int popupToolbarBackground;
-	private int popupToolbarBackgroundColor;
-//	private int readerStatusBackground;
-//	private int readerToolbarBackground;
-//	private int readerToolbarBackgroundVertical;
-//	private int statusTextColor;
-	private int toolbarButtonAlpha;
-	
-	private InterfaceTheme setToolbarButtonAlpha(int toolbarButtonAlpha) {
-		this.toolbarButtonAlpha = toolbarButtonAlpha;
-		return this;
-	}
-//	private InterfaceTheme setBackgrounds(int browserStatusBackground, int browserToolbarBackground,
-//			int browserToolbarBackgroundVertical, int readerStatusBackground, 
-//			int readerToolbarBackground, int readerToolbarBackgroundVertical,
-//			int statusTextColor, int toolbarButtonAlpha) {
-//		this.browserStatusBackground = browserStatusBackground;
-//		this.browserToolbarBackground = browserToolbarBackground;
-//		this.browserToolbarBackgroundVertical = browserToolbarBackgroundVertical;
-//		this.readerStatusBackground = readerStatusBackground;
-//		this.readerToolbarBackground = readerToolbarBackground;
-//		this.readerToolbarBackgroundVertical = readerToolbarBackgroundVertical;
-//		this.statusTextColor = statusTextColor;
-//		this.toolbarButtonAlpha = toolbarButtonAlpha;
-//		return this;
-//	}
+	private final Visuals visuals;
 
-	private InterfaceTheme setBackgrounds(int browserStatusBackground,
-			int browserToolbarBackground, int browserToolbarBackgroundVertical, 
-			int popupToolbarBackground, int browserToolbarBackgroundColor) {
-		this.browserStatusBackground = browserStatusBackground;
-		this.browserToolbarBackground = browserToolbarBackground;
-		this.browserToolbarBackgroundVertical = browserToolbarBackgroundVertical;
-		this.popupToolbarBackground = popupToolbarBackground;
-		this.popupToolbarBackgroundColor = browserToolbarBackgroundColor;
-		return this;
+	static InterfaceTheme create(
+			String code,
+			int themeId,
+			int dialogThemeId,
+			int fsDialogThemeId,
+			int displayNameResourceId,
+			int actionBarBackgroundColorReading,
+			int rootDelimiterResourceId,
+			int rootDelimiterHeight,
+			int browserStatusBackground,
+			int browserToolbarBackground,
+			int browserToolbarBackgroundVertical,
+			int popupToolbarBackground,
+			int popupToolbarBackgroundColor,
+			int toolbarButtonAlpha) {
+		return new InterfaceTheme(
+				code,
+				themeId,
+				dialogThemeId,
+				fsDialogThemeId,
+				displayNameResourceId,
+				actionBarBackgroundColorReading,
+				new Visuals(
+						rootDelimiterResourceId,
+						rootDelimiterHeight,
+						browserStatusBackground,
+						browserToolbarBackground,
+						browserToolbarBackgroundVertical,
+						popupToolbarBackground,
+						popupToolbarBackgroundColor,
+						toolbarButtonAlpha));
 	}
-	
-	private InterfaceTheme(String code, int themeId, int dialogThemeId, int fsDialogThemeId, int displayNameResourceId, int actionBarBackgroundColorReading) {
+
+	private InterfaceTheme(
+			String code,
+			int themeId,
+			int dialogThemeId,
+			int fsDialogThemeId,
+			int displayNameResourceId,
+			int actionBarBackgroundColorReading,
+			Visuals visuals) {
 		this.code = code;
 		this.themeId = themeId;
 		this.dialogThemeId = dialogThemeId;
 		this.fsDialogThemeId = fsDialogThemeId;
 		this.displayNameResourceId = displayNameResourceId;
 		this.actionBarBackgroundColorReading = actionBarBackgroundColorReading;
-	}
-	
-	private InterfaceTheme setRootDelimiter(int resourceId, int height) {
-		this.rootDelimiterResourceId = resourceId;
-		this.rootDelimiterHeight = height;
-		return this;
+		this.visuals = visuals;
 	}
 	
 	@Override
@@ -269,5 +162,36 @@ public class InterfaceTheme {
 
 	public final void applyActionIcons() {
 
+	}
+
+	private static final class Visuals {
+		private final int rootDelimiterResourceId;
+		private final int rootDelimiterHeight;
+		private final int browserStatusBackground;
+		private final int browserToolbarBackground;
+		private final int browserToolbarBackgroundVertical;
+		private final int popupToolbarBackground;
+		private final int popupToolbarBackgroundColor;
+		private final int toolbarButtonAlpha;
+
+		private Visuals(
+				int rootDelimiterResourceId,
+				int rootDelimiterHeight,
+				int browserStatusBackground,
+				int browserToolbarBackground,
+				int browserToolbarBackgroundVertical,
+				int popupToolbarBackground,
+				int popupToolbarBackgroundColor,
+				int toolbarButtonAlpha) {
+			this.rootDelimiterResourceId = rootDelimiterResourceId;
+			this.rootDelimiterHeight = rootDelimiterHeight;
+			this.browserStatusBackground = browserStatusBackground;
+			this.browserToolbarBackground = browserToolbarBackground;
+			this.browserToolbarBackgroundVertical =
+					browserToolbarBackgroundVertical;
+			this.popupToolbarBackground = popupToolbarBackground;
+			this.popupToolbarBackgroundColor = popupToolbarBackgroundColor;
+			this.toolbarButtonAlpha = toolbarButtonAlpha;
+		}
 	}
 }